port(Questionable): TC API12 source (based on TC_forward @ 24d02186)

+++ Questionable/Controller/Utils/PartyWatchDog.cs
@@ -28,7 +28,7 @@ internal sealed class PartyWatchDog : IDisposable
 
     public void Dispose() => _clientState.TerritoryChanged -= TerritoryChanged;
 
-    private unsafe void TerritoryChanged(uint newTerritoryId)
+    private unsafe void TerritoryChanged(ushort newTerritoryId)
     {
         TerritoryIntendedUseEnum intendedUse = (TerritoryIntendedUseEnum)GameMain.Instance()->CurrentTerritoryIntendedUseId;
         switch (intendedUse)
